+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## (Unreleased)
 
 ### Monument
+- (#54) Don't bother freeing memory allocated during the search algorithm.  This makes Monument
+    terminate instantly, often shaving 10s of seconds from the search time.
 - (#53) Add limit on graph size.  Set with `--graph-size-limit`, defaults to 100K chunks.
 - (#50) Add `bobs_only` and `singles_only`.
 - (#50) Fix some dead links in Monument's guide.